made encoder virtual class
