+++ j2/platformio.ini
@@ -9,7 +9,7 @@
 ; https://docs.platformio.org/page/projectconf.html
 
 [env:esp32-s3-devkitc-1]
-platform = espressif32
+platform = https://github.com/platformio/platform-espressif32.git
 board = esp32-s3-devkitc-1
 framework = arduino
 
@@ -19,6 +19,10 @@ monitor_raw = true
 lib_deps = 
     rover_can_lib=symlink://../lib/rover_can_lib
     lib_rover=symlink://../lib/lib_rover
+
+build_unflags =
+    -w CCFLAGS
+    -std=gnu++11
 
 build_flags =
     '-D ARDUINO_USB_CDC_ON_BOOT=1'
